// mobile/lib/types.ts
export interface Address {
  id: string;
  name: string;
  phone: string;
  street: string;
  city: string;
  county: string | null;
  postalCode: string | null;
  country?: string;
  isDefault: boolean;
}
// ...
export interface CheckoutRequest {
  cartId: string;
  currency: string;
  amount: number;
  paymentMethod: "card" | "mobile_money";
  mobileMoney?: {
    network: "MPESA" | "AIRTEL" | "MTN";
    phone: string;
  };
  customer: {
    name: string;
    email: string;
  };
  discreet?: boolean;
  shippingAddress?: string;
  couponCode?: string;
  affiliateCode?: string;
}